Moving all logic to the root
Fix https://github.com/remix-run/remix/pull/1428#pullrequestreview-847201337
