+++ examples/gdpr-cookie-consent/README.md
@@ -12,10 +12,12 @@ Open this example on [CodeSandbox](https://codesandbox.com):
 
 ## Example
 
-Users will be presented with a GDPR consent form on every page ([app/routes/index.tsx](app/routes/index.tsx)) till they submit the accept button.
-Once they submit the consent form, a dummy tracking [scripts](public/dummy-analytics-script.js) at the [app/root.tsx](app/root.tsx) will start tracking the user`s data (open the browser console too see the dummy tracking message).
+Users will be presented with a GDPR consent form on every page [app/root.tsx](app/root.tsx) till they submit the accept button.
+Once they submit the consent form, a dummy tracking [scripts](public/dummy-analytics-script.js) at the [app/root.tsx](app/root.tsx)  will start tracking the user`s data (open the browser console too see the dummy tracking message).
 
-The example is using Remix Cookie API
+> If you want to reset the example delete the `gdpr-consent` cookie in the `Application`/`cookies` in the browser's developer tools.
+
+The example is using [Remix Cookie API.]([https://link](https://remix.run/docs/en/v1/api/remix#cookies)) 
 
 ## Related Links
